server/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
protobuf>=3.19.5,<5.0.0
python-dotenv==1.0.1
Werkzeug==3.0.2
gunicorn==20.1.0
google-generativeai==0.8.6
requests==2.31.0